Updated database.sample.yml switching to mysql2 adapter.
Also removed some comments that appeared obsolete and directed developers to the
README file instead, which though probably still oudated might be more relevant.
I left in the staging, integration and production databases but we could perhaps
do with more info on their use as the sample config does not look realistic.