WEB-4944 Quick fix to prevent empty attribute lists in data search.
Needs a better solution but for now attributes are reset, clade is cleared and
flash notice is added if user arrives on data search, or searches, with a clade
that has no data. For users that come up against a reduced set of attributes
when trying to modify their search there is now a new search link that will
reload data search with no parameters.