WEB-5128 - Upgrade paperclip
Also fixes a typo with the configuration variable I'd created for
attachments. Collections are still the only model that allows you to
upload thumbnails in non-prod environments. Wrote a few more tests for
things, also fixed a problem with preloading the file_name for
collections from the Taxon class (it's now needed to check whether the
logo is local).