WEB-4985 - sort collection items with blanks last
Interesting problem: the SolrAPI.text_filter(sort_field) puts empty
strings in if fields are missing. For THIS bug, I've simply altered the
method that uses it to exclude the field if it's blank (thus allowing
the sortMissingLast attribute to work), but perhaps we need to alter the
.text_filter method. I'm not sure; will talk to PL about it. It seems
risky.  ...Perhaps we just need a second method that returns nil instead
of an empty string. ...but I've not tested that it would work with nil
(I'm just not adding it to the hash at all.)