Removing FB/Twitter (take 3)
This time I'm leaving all of the views/partials, since that's where the
metadata gets defined (and I didn't want to *always* define it, that
would be inefficient).

I still removed that Facebook controller, because it's not clear what
it is, and I cannot find any hints with Google searches.