WEB-5156 Added admin tool for collection item recount.
I used this as an excuse to make the admin UI a *tiny* bit better (more
general ajax usage; this has less code than before, but a little more of
a framework in the view so it's not *quite* as ugly), and added to it
the ability to re-count collection items.  Thus, both of the problems
explained in the ticket can be solved from within the UI, now: the admin
page's re-count, and the re-index on the collection page.