Added PageMeta and related classes
Ruby On Rails style guide suggests that no more than 2
instance variables should be passed to a view. Trying to
cope with this we are creating PageMeta class which would
be a wrapper for all specific classes from various controllers
Current convention is that we call specific class by the name
of restful method -- for example IndexMeta takes care of
index method. If class covers many methods we call it Meta